Initial commit of local progress
This commit marks the latest checkpoint in our development. Completed features include the timeline page with all events for each match and the home page, which includes a basic filter to display videos for specific events (to be enhanced in the future). Moving forward, the next steps will focus on improving the design and implementing additional features to enhance functionality.