Updated SDK to 53
This commit include the update of the Expo Software Development Kit to 53. This also means updating the react, react-native and expo versions. NOTE: As this this commit, React Native Testing Library RNTL, DOES NOT SUPPORT React v19. Which means it' not possible to run the tests on the new SDK 53 yet.